requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
distlib==0.3.6
filelock==3.9.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector-python==8.0.32
numpy==1.24.2
opencv-python==4.7.0.68
outcome==1.2.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==2.6.2
protobuf==3.20.3
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytweening==1.0.4
pytz==2022.7.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
virtualenv==20.17.1
wsproto==1.2.0
yarl==1.8.2
